server/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.4
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
attrs==21.4.0
autopep8==1.6.0
azure-functions==1.21.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.1.0
certifi==2024.2.2
cffi==1.15.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.4
comm==0.2.2
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==3.2
dnspython==2.7.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.15.3
Flask==1.1.2
fonttools==4.33.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.8.2
joblib==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-console==6.4.3
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaggle==1.6.14
kiwisolver==1.4.3
langchain-core==0.3.29
langchain-text-splitters==0.3.4
langsmith==0.2.10
loguru==0.7.3
MarkupSafe==2.1.1
marshmallow==3.23.3
matplotlib==3.5.2
matplotlib-inline==0.1.7
mistune==0.8.4
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.6.0
notebook==6.4.12
numexpr==2.10.2
numpy==1.26.4
openai==0.27.0
openapi-schema-pydantic==1.2.4
orjson==3.10.13
outcome==1.1.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.1.1
pinecone-client==2.2.1
platformdirs==4.2.2
plotly==5.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.43
propcache==0.2.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.18.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
qtconsole==5.3.1
QtPy==2.1.0
requests==2.32.2
requests-toolbelt==1.0.0
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.18.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.36
sqlparse==0.4.1
stack-data==0.6.3
starlette==0.27.0
tenacity==8.5.0
terminado==0.15.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
torch==1.12.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.20.0
trio-websocket==0.9.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.22.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wsproto==1.0.0
yarl==1.18.3
zipp==3.18.2
